extract max-quota query from Panel instead move it to the maxQuota hook. Use it if no custom query is provided by the caller.
