+++ src/components/domain/DomainManager.js
@@ -20,7 +20,7 @@ import { clusterStore, clusterStoreActions } from "../StoreProvider";
 import ProjectsPerDomain from "./ProjectsPerDomain";
 
 const DomainManager = (props) => {
-  const { serviceType, currentCategory, currentResource, currentAZ, subRoute, setMaxQuota, mergeOps } = props;
+  const { serviceType, currentCategory, currentResource, currentAZ, subRoute, mergeOps } = props;
   const { domainData } = clusterStore();
   const { setDomainData } = clusterStoreActions();
   const domainQueryResult = useQuery({ queryKey: ["domains", "", ""] });
@@ -43,7 +43,6 @@ const DomainManager = (props) => {
         currentCategory={currentCategory}
         currentAZ={currentAZ}
         subRoute={subRoute}
-        setMaxQuota={setMaxQuota}
         mergeOps={mergeOps}
       />
     )
