+++ src/components/mainView/subComponents/MaxQuota.js
@@ -21,8 +21,9 @@ import { Unit, valueWithUnit } from "../../../lib/unit";
 const MaxQuota = (props) => {
   const { project, resource, serviceType, postMaxQuota } = props;
   const { isPanelView = false } = props;
-  const maxQuota = resource?.max_quota;
+  const maxQuotaValue = resource?.max_quota;
   const unit = new Unit(resource.unit);
+
   const { MaxQuotaInput, maxQuotaInputProps, MaxQuotaEdit, maxQuotaEditProps } = useMaxQuotaSets({
     project: project,
     resource: resource,
@@ -37,7 +38,7 @@ const MaxQuota = (props) => {
       <MaxQuotaEdit {...maxQuotaEditProps} iconOnlyView={true} subduedView={true} />
     </span>
   ) : (
-    <span>Max Quota: {valueWithUnit(maxQuota, unit)}</span>
+    <span>Max Quota: {valueWithUnit(maxQuotaValue, unit)}</span>
   );
 };
