+++ src/components/project/ProjectManager.js
@@ -22,7 +22,7 @@ import { globalStoreActions, domainStoreActions, domainStore } from "../StorePro
 import { LoadingIndicator } from "@cloudoperators/juno-ui-components";
 
 const ProjectManager = (props) => {
-  const { serviceType, currentCategory, currentResource, currentAZ, subRoute, setMaxQuota, mergeOps } = props;
+  const { serviceType, currentCategory, currentResource, currentAZ, subRoute, mergeOps } = props;
   const resourceName = currentResource.name;
   const { refetchProjectAPI } = projectStore();
   const { setRefetchProjectAPI } = projectStoreActions();
@@ -61,7 +61,6 @@ const ProjectManager = (props) => {
       currentAZ={currentAZ}
       projects={projects}
       subRoute={subRoute}
-      setMaxQuota={setMaxQuota}
       mergeOps={mergeOps}
     />
   ) : (
